fixed the modal URL targeting if hash name includes a dash
Had to add this indexOf to look for a "-" because we are also finding people's bios by ID on the people pages and it was acting like a weird modal since those URLs had hashes as well. Now, if there is a "-" which should be in people's names and not in the contactModal URLs, it won't do the modal thing. It only does it if there is a # and no "-" in the URL.